:recycle: root layout을 app layer로 이동 #1
