+++ app/layout.tsx
@@ -1,20 +1,1 @@
-import '@/styles/global.scss';
-import { ReactNode } from 'react';
-
-export const metadata = {
-  title: 'CAMUS',
-  // description: 'Generated by Next.js',
-  icons: {
-    icon: '/images/logo.svg',
-  },
-};
-
-const RootLayout = ({ children }: { children: ReactNode }) => {
-  return (
-    <html lang='en'>
-      <body>{children}</body>
-    </html>
-  );
-};
-
-export default RootLayout;
+export { default, metadata } from '@/app/layouts/root';
