+++ src/app/layouts/root.tsx
@@ -1,0 +1,20 @@
+import '@/styles/global.scss';
+import type { ReactNode } from 'react';
+
+export const metadata = {
+  title: 'CAMUS',
+  // description: 'Generated by Next.js',
+  icons: {
+    icon: '/images/logo.svg',
+  },
+};
+
+const RootLayout = ({ children }: { children: ReactNode }) => {
+  return (
+    <html lang='en'>
+      <body>{children}</body>
+    </html>
+  );
+};
+
+export default RootLayout;
